(PC-29315)[API] refactor: move VenueTypeCode definition to offerer models
